Move of repository from internal server to GitHub
Date:   Tue Mar 28 10:13:39 2017 +0200
Exclude JNA and JNA-Platform from Maven dependency list.
This is to cut a dependency on a different license type.

Date:   Thu Mar 9 16:37:23 2017 +0100
Update README.md to provide comprehensive information on project usage.

Date:   Thu Mar 2 17:40:09 2017 +0100
Prepare code for release-readiness.
+ added license.txt
+ added license headers to all source files
+ set version in POM file to 1.0.0.

Date:   Wed Dec 21 17:28:07 2016 +0100
Implement default handling if -Dtestcontext.implclassname is missing.
Now try to find the impl class for the given interface in the same
package by taking the interface name and postfixing "Impl".
That way customers don't have to implement that system property.

Date:   Thu Nov 24 16:24:43 2016 +0100
We need to be able to explicitly set the proxy server url for Jenkins.
A BaseTestContext interface method will allow to retrieve the proxy
server url.

Date:   Tue Nov 22 18:40:04 2016 +0100
Updated instructions how Maven can use this project as dependency.
Date:   Mon Nov 7 16:50:51 2016 +0100
Initial checkin of cqe-test-container project.