Allow to toggle between WebDriver 2.x or 3.x.
Build two different jar files, one which depends on WebDriver 3.x or
newer, and one which depends on WebDriver 2.x. If a consuming project
needs WebDriver 2, then look at README.md for instructions how to use
Test Drop-in Framework with a dependency on that version of WebDriver.

The resulting jar files have pom files with different artifact ID's.