Fix for vulnerability CVE-2018-7489.
FasterXML jackson-databind before 2.8.11.1 and 2.9.x before 2.9.5 allows
unauthenticated remote code execution because of an incomplete fix for
the CVE-2017-7525 deserialization flaw. This is exploitable by sending
maliciously crafted JSON input to the readValue method of the
ObjectMapper, bypassing a blacklist that is ineffective if the c3p0
libraries are available in the classpath.